Authorization, jwt, passport strategy

+++ src/tasks/tasks.service.ts
@@ -5,6 +5,7 @@ import { Task } from './task.entity';
 import { CreateTaskDto } from './dto/create-task.dto';
 import { TaskStatus } from './task-status.enum';
 import { GetTasFilterDto } from './dto/get-task-filter.dto';
+import { User } from 'src/auth/user.entity';
 
 @Injectable()
 export class TasksService {
@@ -14,13 +15,13 @@ export class TasksService {
         private taskRepository: TaskRepository
     ) { };
     
-    async getTasks(filterDto:GetTasFilterDto): Promise<Task[]> { 
-        return await this.taskRepository.getTasks(filterDto);
+    async getTasks(filterDto:GetTasFilterDto, user:User): Promise<Task[]> { 
+        return await this.taskRepository.getTasks(filterDto, user);
      };
  
 
-    async getTaskById(id: number) :Promise<Task> { 
-        const found = await this.taskRepository.findOne(id);
+    async getTaskById(id: number,user:User) :Promise<Task> { 
+        const found = await this.taskRepository.findOne({where: {id, userId: user.id }});
         if(!found) { 
             throw new NotFoundException(`Task with id: "${id}" not found`);
         };
@@ -28,20 +29,20 @@ export class TasksService {
     };
 
    
-    async createTask(task:CreateTaskDto): Promise<Task> { 
-        const newTask = await this.taskRepository.createTask(task);
+    async createTask(task:CreateTaskDto, user:User): Promise<Task> { 
+        const newTask = await this.taskRepository.createTask(task, user);
         return newTask;
     };
 
-    async updateTaskStatus(id:number, status:TaskStatus):Promise<Task> {
-        const task = await this.getTaskById(id);
+    async updateTaskStatus(id:number, status:TaskStatus, user:User):Promise<Task> {
+        const task = await this.getTaskById(id, user);
         task.status = status;
         await task.save();
         return task;
     };
  
-    async deleteTask(id:number): Promise<void> { 
-      const result = await this.taskRepository.delete({id});
+    async deleteTask(id:number, user:User): Promise<void> { 
+      const result = await this.taskRepository.delete({id, userId: user.id});
       if(result.affected === 0) { 
         throw new NotFoundException(`Task with id: ${id} not found!`); 
       };
